app/requirements.txt
fastapi==0.92.0
pydantic==1.10.4
uvicorn==0.20.0
python-dotenv
sqlalchemy
pillow
mysql-connector-python
celery[redis]